requirements.txt
matplotlib==3.7.5
numpy==1.24.3
pandas==2.0.3
scikit_learn==1.3.0
scipy==1.10.1
spams_bin==2.6.10
tqdm==4.67.1
